+++ src/test/php/web/rest/unittest/People.class.php
@@ -1,0 +1,13 @@
+<?php namespace web\rest\unittest;
+
+class People {
+
+  /** @var web.rest.unittest.Person[] */
+  private $list;
+
+  public function __construct(... $list) {
+    $this->list= $list;
+  }
+
+  public function all() { return $this->list; }
+}